+++ src/irmin-pack/IO.mli
@@ -15,3 +15,4 @@
  *)
 
 include IO_intf.IO
+(** @inline *)