+++ src/irmin/export_for_backends.ml
@@ -1,5 +1,5 @@
 (*
- * Copyright (c) 2013-2020 Thomas Gazagnaire <thomas@gazagnaire.org>
+ * Copyright (c) 2021 Craig Ferguson <craig@tarides.com>
  *
  * Permission to use, copy, modify, and distribute this software for any
  * purpose with or without fee is hereby granted, provided that the above
@@ -7,36 +7,15 @@
  *
  * THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES
  * WITH REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF
- * MERCHANTABILITY AND FITNESIrmin. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR
+ * MERCHANTABILITY AND FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR
  * ANY SPECIAL, DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
  * WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS, WHETHER IN AN
  * ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
  * OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
  *)
 
-open! Import
-open Store_properties
+type read = Perms.read
+type write = Perms.write
+type read_write = Perms.read_write
 
-module type S = sig
-  type t
-
-  val v : string -> t Lwt.t
-
-  include CLOSEABLE with type _ t := t
-
-  val read_flip : t -> bool Lwt.t
-  val write_flip : bool -> t -> unit Lwt.t
-end
-
-module IO : S
-
-module Lock : sig
-  type t
-
-  val v : string -> t Lwt.t
-
-  include CLOSEABLE with type _ t := t
-
-  val unlink : string -> unit Lwt.t
-  val test : string -> bool
-end
+module Store_properties = S.Store_properties